+++ src/routes/authRoutes.ts
@@ -9,13 +9,24 @@ import {
   getUser,
 } from "../controllers/authController";
 import { errorCatch } from "../utils/error/errorCatch";
+import { validateData } from "../middleware/zodValidation";
+import {
+  loginSchema,
+  otpSchema,
+  registerSchema,
+  resendOtpSchema,
+} from "../utils/zodSchemas";
 
 const router = express.Router();
 
-router.post("/sendotp", errorCatch(registerWithEmail));
-router.post("/validate", errorCatch(verifyEmailOtp));
-router.post("/login", errorCatch(login));
-router.post("/resentotp", errorCatch(resendOtp));
+router.post(
+  "/sendotp",
+  validateData(registerSchema),
+  errorCatch(registerWithEmail)
+);
+router.post("/validate", validateData(otpSchema), errorCatch(verifyEmailOtp));
+router.post("/login", validateData(loginSchema), errorCatch(login));
+router.post("/resentotp", validateData(resendOtpSchema), errorCatch(resendOtp));
 router.post("/refresh", errorCatch(refreshToken));
 router.post("/logout", errorCatch(logout));
 router.get("/me", errorCatch(getUser));
